+++ tests/unit/core/factories.py
@@ -63,9 +63,7 @@ class CuratedListPageFactory(wagtail_factories.PageFactory):
     topics = wagtail_factories.StreamFieldFactory(
         {
             'title': wagtail_factories.CharBlockFactory,
-            'pages': wagtail_factories.ListBlockFactory(
-                wagtail_factories.PageFactory
-            ),
+            # lessons_and_placeholders need to be added not via the factory, for now...
         }
     )
 
@@ -140,9 +138,10 @@ class SimpleVideoBlockFactory(wagtail_factories.StructBlockFactory):
         model = blocks.SimpleVideoBlock
 
 
-class CuratedTopicBlockfactory(wagtail_factories.StructBlockFactory):
+class CuratedTopicBlockFactory(wagtail_factories.StructBlockFactory):
     title = factory.fuzzy.FuzzyText(length=255)
-    pages = factory.SubFactory(DetailPageFactory)
+    # lessons_and_placeholders need to be added via a helper, not via the factory, for
+    # now - see tests.helpers.add_lessons_and_placeholders_to_curated_list_page
 
     class Meta:
         model = blocks.CuratedTopicBlock